Fixed TextureArrays

+++ res/shaders/basic.vert
@@ -3,10 +3,10 @@
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec3 aNormal;
 layout (location = 2) in vec2 aTexCoord;
-layout (location = 3) in int aTextureIndex;
+layout (location = 3) in float aTextureIndex;
 
 out vec2 TexCoord;
-flat out int TextureIndex;
+flat out float TextureIndex;
 
 uniform mat4 model;
 uniform mat4 view;
